fix(core): layout grid examples
