+++ apps/docs/src/app/core/component-docs/layout-grid/examples/index.ts
@@ -1,14 +1,14 @@
 import { ExampleLayoutGridBlockComponent } from './example-layout-grid-block.component';
+import { LayoutGridBasicExampleComponent } from './layout-grid-basic-example.component';
 import {
-    LayoutGridBasicExampleComponent,
     LayoutGridNestingExampleComponent,
     LayoutGridNoGapExampleComponent,
     LayoutGridOffsetExampleComponent,
     LayoutGridResponsiveExampleComponent,
     LayoutGridResponsiveOffsetExample,
-    LayoutGridRowExampleComponent,
-    LayoutGridGrowingExampleComponent
+    LayoutGridRowExampleComponent
 } from './layout-grid-examples.component';
+import { LayoutGridGrowingExampleComponent } from './layout-grid-growing-example.component';
 
 export * from './layout-grid-examples.component';
 export * from './example-layout-grid-block.component';
