+++ apps/docs/src/app/core/component-docs/layout-grid/examples/layout-grid-basic-example.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-layout-grid-basic-example',
+    templateUrl: './layout-grid-basic-example.component.html'
+})
+export class LayoutGridBasicExampleComponent {}
