+++ apps/docs/src/app/core/component-docs/layout-grid/examples/layout-grid-examples.component.ts
@@ -1,12 +1,6 @@
 import { Component } from '@angular/core';
 
 @Component({
-    selector: 'fd-layout-grid-basic-example',
-    templateUrl: './layout-grid-basic-example.component.html'
-})
-export class LayoutGridBasicExampleComponent {}
-
-@Component({
     selector: 'fd-layout-grid-no-gap-example',
     templateUrl: './layout-grid-no-gap-example.component.html'
 })
@@ -17,12 +11,6 @@ export class LayoutGridNoGapExampleComponent {}
     templateUrl: './layout-grid-offset-example.component.html'
 })
 export class LayoutGridOffsetExampleComponent {}
-
-@Component({
-    selector: 'fd-layout-grid-growing-example',
-    templateUrl: './layout-grid-growing-example.component.html'
-})
-export class LayoutGridGrowingExampleComponent {}
 
 @Component({
     selector: 'fd-layout-grid-responsive-example',
