+++ apps/docs/src/app/core/component-docs/layout-grid/examples/layout-grid-growing-example.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-layout-grid-growing-example',
+    templateUrl: './layout-grid-growing-example.component.html'
+})
+export class LayoutGridGrowingExampleComponent {}
